// frontend/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = T | null | undefined;
// ...
export type BazelInvocation = Node & {
  __typename?: 'BazelInvocation';
  actions?: Maybe<Array<Action>>;
  authenticatedUser?: Maybe<AuthenticatedUser>;
  bazelVersion?: Maybe<Scalars['String']['output']>;
  bepCompleted: Scalars['Boolean']['output'];
  build?: Maybe<Build>;
  /** JSON representation of the canonical command line options. */
  canonicalCommandLine?: Maybe<Scalars['Map']['output']>;
  configurations?: Maybe<Array<Configuration>>;
  connectionMetadata?: Maybe<ConnectionMetadata>;
  endedAt?: Maybe<Scalars['Time']['output']>;
  environmentVariables?: Maybe<Scalars['Map']['output']>;
  exitCodeCode?: Maybe<Scalars['Int']['output']>;
  exitCodeName?: Maybe<Scalars['String']['output']>;
  hostname?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  instanceName: InstanceName;
  invocationFiles?: Maybe<Array<InvocationFiles>>;
  invocationID: Scalars['UUID']['output'];
  invocationTargets: InvocationTargetConnection;
  metrics?: Maybe<Metrics>;
  numFetches?: Maybe<Scalars['Int']['output']>;
  /** JSON representation of the parsed command line options */
  optionsParsed?: Maybe<Scalars['Map']['output']>;
  /** JSON representation of the original command line options. */
  originalCommandLine?: Maybe<Scalars['Map']['output']>;
  profile?: Maybe<Profile>;
  sourceControl?: Maybe<Array<SourceControl>>;
  startedAt?: Maybe<Scalars['Time']['output']>;
  tags: InvocationTagConnection;
  username?: Maybe<Scalars['String']['output']>;
};
// ...
export type InstanceName = Node & {
  __typename?: 'InstanceName';
  bazelInvocations?: Maybe<Array<BazelInvocation>>;
  builds?: Maybe<Array<Build>>;
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  targets?: Maybe<Array<Target>>;
};
// ...
export type InvocationTarget = Node & {
  __typename?: 'InvocationTarget';
  abortReason: InvocationTargetAbortReason;
  bazelInvocation: BazelInvocation;
  configuration?: Maybe<Configuration>;
  durationInMs?: Maybe<Scalars['Int']['output']>;
  endTimeInMs?: Maybe<Scalars['Int']['output']>;
  failureMessage?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  startTimeInMs?: Maybe<Scalars['Int']['output']>;
  success: Scalars['Boolean']['output'];
  tags?: Maybe<Array<Scalars['String']['output']>>;
  target: Target;
  targetFiles?: Maybe<Array<InvocationFiles>>;
  testSummary?: Maybe<Array<TestSummary>>;
};
// ...
export enum OrderDirection {
  /** Specifies an ascending order for a given `orderBy` argument. */
  Asc = 'ASC',
  /** Specifies a descending order for a given `orderBy` argument. */
  Desc = 'DESC'
}

/**
 * Information about pagination in a connection.
 * https://relay.dev/graphql/connections.htm#sec-undefined.PageInfo
 */
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['Cursor']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['Cursor']['output']>;
};

export type Profile = {
  __typename?: 'Profile';
  digest: Scalars['String']['output'];
  digestFunction: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  sizeInBytes: Scalars['Int']['output'];
};

export type Query = {
  __typename?: 'Query';
  findBazelInvocations: BazelInvocationConnection;
  findBuilds: BuildConnection;
  findTargets: TargetConnection;
  findTestSummaries: TestSummaryConnection;
  getAuthenticatedUser?: Maybe<AuthenticatedUser>;
  getBazelInvocation?: Maybe<BazelInvocation>;
  getBuild?: Maybe<Build>;
  getTarget?: Maybe<Target>;
  /** Fetches an object given its ID. */
  node?: Maybe<Node>;
  /** Lookup nodes by a list of IDs. */
  nodes: Array<Maybe<Node>>;
};
// ...
export type TestResult = Node & {
  __typename?: 'TestResult';
  attempt: Scalars['Int']['output'];
  cachedLocally?: Maybe<Scalars['Boolean']['output']>;
  cachedRemotely?: Maybe<Scalars['Boolean']['output']>;
  exitCode?: Maybe<Scalars['Int']['output']>;
  hostname?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  run: Scalars['Int']['output'];
  shard: Scalars['Int']['output'];
  status?: Maybe<Scalars['String']['output']>;
  statusDetails?: Maybe<Scalars['String']['output']>;
  strategy?: Maybe<Scalars['String']['output']>;
  testAttemptDurationInMs?: Maybe<Scalars['Int']['output']>;
  testAttemptStart?: Maybe<Scalars['Time']['output']>;
  testResultFiles?: Maybe<Array<TestResultFile>>;
  testSummary: TestSummary;
  timingBreakdown?: Maybe<Scalars['Map']['output']>;
  warning?: Maybe<Array<Scalars['String']['output']>>;
};
// ...
export type TestSummary = Node & {
  __typename?: 'TestSummary';
  attemptCount?: Maybe<Scalars['Int']['output']>;
  firstStartTime?: Maybe<Scalars['Time']['output']>;
  id: Scalars['ID']['output'];
  invocationTarget: InvocationTarget;
  lastStopTime?: Maybe<Scalars['Time']['output']>;
  overallStatus?: Maybe<Scalars['String']['output']>;
  runCount?: Maybe<Scalars['Int']['output']>;
  shardCount?: Maybe<Scalars['Int']['output']>;
  testResults?: Maybe<Array<TestResult>>;
  totalNumCached?: Maybe<Scalars['Int']['output']>;
  totalRunCount?: Maybe<Scalars['Int']['output']>;
  totalRunDurationInMs?: Maybe<Scalars['Int']['output']>;
};

/** A connection to a list of items. */
export type TestSummaryConnection = {
  __typename?: 'TestSummaryConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<TestSummaryEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type TestSummaryEdge = {
  __typename?: 'TestSummaryEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<TestSummary>;
};

/** Ordering options for TestSummary connections */
export type TestSummaryOrder = {
  /** The ordering direction. */
  direction?: OrderDirection;
  /** The field by which to order TestSummaries. */
  field: TestSummaryOrderField;
};

/** Properties by which TestSummary connections can be ordered. */
export enum TestSummaryOrderField {
  FirstStartTime = 'FIRST_START_TIME',
  TotalRunDurationInMs = 'TOTAL_RUN_DURATION_IN_MS'
}

/**
 * TestSummaryWhereInput is used for filtering TestSummary objects.
 * Input was generated by ent.
 */
export type TestSummaryWhereInput = {
// ...
  overallStatusIn?: InputMaybe<Array<Scalars['String']['input']>>;
// ...
};
// ...
export type GetTestsForInvocationQueryVariables = Exact<{
  after?: InputMaybe<Scalars['Cursor']['input']>;
  first?: InputMaybe<Scalars['Int']['input']>;
  before?: InputMaybe<Scalars['Cursor']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
  orderBy?: InputMaybe<TestSummaryOrder>;
  where?: InputMaybe<TestSummaryWhereInput>;
}>;


export type GetTestsForInvocationQuery = { __typename?: 'Query', findTestSummaries: { __typename?: 'TestSummaryConnection', pageInfo: { __typename?: 'PageInfo', startCursor?: any | null, endCursor?: any | null, hasNextPage: boolean, hasPreviousPage: boolean }, edges?: Array<{ __typename?: 'TestSummaryEdge', node?: { __typename?: 'TestSummary', id: string, overallStatus?: string | null, totalRunDurationInMs?: number | null, runCount?: number | null, attemptCount?: number | null, shardCount?: number | null, totalNumCached?: number | null, firstStartTime?: any | null, lastStopTime?: any | null, testResults?: Array<{ __typename?: 'TestResult', cachedLocally?: boolean | null, cachedRemotely?: boolean | null }> | null, invocationTarget: { __typename?: 'InvocationTarget', bazelInvocation: { __typename?: 'BazelInvocation', invocationID: any }, target: { __typename?: 'Target', id: string, label: string, aspect: string, targetKind: string, instanceName: { __typename?: 'InstanceName', name: string } } } } | null } | null> | null } };
// ...
export const GetTestsForInvocationDocument = {"__meta__":{"hash":"00e5104f456408202736756c11e1cd8dd9e62047"},"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetTestsForInvocation"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"after"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Cursor"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"first"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"before"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Cursor"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"last"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"orderBy"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"TestSummaryOrder"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"where"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"TestSummaryWhereInput"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"findTestSummaries"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"after"},"value":{"kind":"Variable","name":{"kind":"Name","value":"after"}}},{"kind":"Argument","name":{"kind":"Name","value":"first"},"value":{"kind":"Variable","name":{"kind":"Name","value":"first"}}},{"kind":"Argument","name":{"kind":"Name","value":"before"},"value":{"kind":"Variable","name":{"kind":"Name","value":"before"}}},{"kind":"Argument","name":{"kind":"Name","value":"last"},"value":{"kind":"Variable","name":{"kind":"Name","value":"last"}}},{"kind":"Argument","name":{"kind":"Name","value":"orderBy"},"value":{"kind":"Variable","name":{"kind":"Name","value":"orderBy"}}},{"kind":"Argument","name":{"kind":"Name","value":"where"},"value":{"kind":"Variable","name":{"kind":"Name","value":"where"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"pageInfo"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"startCursor"}},{"kind":"Field","name":{"kind":"Name","value":"endCursor"}},{"kind":"Field","name":{"kind":"Name","value":"hasNextPage"}},{"kind":"Field","name":{"kind":"Name","value":"hasPreviousPage"}}]}},{"kind":"Field","name":{"kind":"Name","value":"edges"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"node"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"overallStatus"}},{"kind":"Field","name":{"kind":"Name","value":"totalRunDurationInMs"}},{"kind":"Field","name":{"kind":"Name","value":"runCount"}},{"kind":"Field","name":{"kind":"Name","value":"attemptCount"}},{"kind":"Field","name":{"kind":"Name","value":"shardCount"}},{"kind":"Field","name":{"kind":"Name","value":"totalNumCached"}},{"kind":"Field","name":{"kind":"Name","value":"firstStartTime"}},{"kind":"Field","name":{"kind":"Name","value":"lastStopTime"}},{"kind":"Field","name":{"kind":"Name","value":"testResults"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"cachedLocally"}},{"kind":"Field","name":{"kind":"Name","value":"cachedRemotely"}}]}},{"kind":"Field","name":{"kind":"Name","value":"invocationTarget"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"bazelInvocation"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"invocationID"}}]}},{"kind":"Field","name":{"kind":"Name","value":"target"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"instanceName"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}}]}},{"kind":"Field","name":{"kind":"Name","value":"label"}},{"kind":"Field","name":{"kind":"Name","value":"aspect"}},{"kind":"Field","name":{"kind":"Name","value":"targetKind"}}]}}]}}]}}]}}]}}]}}]} as unknown as DocumentNode<GetTestsForInvocationQuery, GetTestsForInvocationQueryVariables>;